# Conflicts:
#	src/main/java/online/store/controllers/ProductController.java
#	src/main/java/online/store/model/DTO/ProductDTO.java
#	src/main/java/online/store/model/Product.java
#	src/main/java/online/store/model/Role.java
#	src/main/java/online/store/repositories/AuthorRepository.java
#	src/main/java/online/store/repositories/ProductCategoryRepository.java
#	src/main/java/online/store/repositories/ProductRepository.java
#	src/main/java/online/store/services/ProductCategoryService.java
#	src/main/java/online/store/services/ProductService.java

+++ src/main/java/online/store/model/OrderLine.java
@@ -5,7 +5,6 @@ import lombok.*;
 
 import javax.persistence.*;
 import java.io.Serializable;
-import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -19,7 +18,7 @@ public class OrderLine implements Serializable {
     private Long id;
 
     private Integer quantity;
-    private BigDecimal totalPrice;
+    private Double totalPrice;
 
     @ManyToOne
     @JoinColumn(name = "product_id")
